requirements.txt
asgiref==3.9.1
Django==5.1.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
